+++ GlobalUsings.cs
@@ -29,4 +29,5 @@ global using static WECCL.Utils.MenuUtils;
 global using static WECCL.Content.CustomContent;
 global using static WECCL.Utils.FileNameUtils;
 global using static WECCL.Utils.ArrayUtils;
+global using static WECCL.Utils.LogUtils;
 global using static WECCL.Utils.DataSharing;